Add locale mixins for date and WorkToDo
